requirements.txt
pandas==2.0.3
streamlit==1.25.0
beautifulsoup4==4.12.2
azure-ai-formrecognizer==3.2.1
requests==2.26.0
azure-common==1.1.28
azure-core==1.28.0